integrates\fix(back): #4014.159 sort risk level
- The sort risk level was developed as a float but it is an integer
- The default value for the sort risk level must be -1 according to
the 0135_update_toe_sorts_file_risk migration

+++ integrates/back/src/api/mutations/update_toe_lines_sorts.py
@@ -47,7 +47,7 @@ async def mutate(
     group_name: str,
     root_nickname: str,
     filename: str,
-    sorts_risk_level: float,
+    sorts_risk_level: int,
 ) -> SimplePayloadType:
     try:
         group_roots_loader = info.context.loaders.group_roots
